Add pytest and coverage plugin, update make, update git actions
