+++ .github/workflows/pythonpackage.yml
@@ -26,4 +26,7 @@ jobs:
       run: |
         pip install flake8
         make lint
+    - name: Test and CC
+      run: |
+        make test-n-cover
