+++ Makefile
@@ -21,7 +21,7 @@ PYTHON = python3
 
 # .PHONY defines parts of the makefile that are not dependant on any specific file
 # This is most often used to store functions
-.PHONY = all help setup format lint run_tests debug clean
+.PHONY = all help setup format lint test-n-cover
 
 # Defining an array variable
 FILES = input output
@@ -37,7 +37,7 @@ help:
 	@echo "$(BOLD_CYAN)make setup$(RESET_STYLES) : Setup pyflask-service"
 	@echo "$(BOLD_CYAN)make format$(RESET_STYLES) : Format and fix python code in pyflask-service"
 	@echo "$(BOLD_CYAN)make lint$(RESET_STYLES) : Lint pyflask-service"
-	@echo "$(BOLD_CYAN)make run_tests$(RESET_STYLES) : Test pyflask-service"
+	@echo "$(BOLD_CYAN)make test-n-cover$(RESET_STYLES) : Test and code coverage pyflask-service"
 	@echo "$(BOLD_CYAN)make debug$(RESET_STYLES) : Debug pyflask-service"
 	@echo "$(BOLD_CYAN)make clean$(RESET_STYLES) : Clean pyflask-service"
 	@echo "$(BOLD_CYAN)make dev-run$(RESET_STYLES) : Run pyflask-service in environment=development"
@@ -95,8 +95,8 @@ lint: #: Run static analysis with flake8, radon, mypy and bandit
 	@echo "\n"
 
 
-run_tests: #: Test with pytest
-	@echo "\n$(BOLD_CYAN)Testing with pytest$(RESET_STYLES) ❄️"
+test-n-cover: #: Test with pytest, Code coverage with pytest-cov plugin
+	@echo "\n$(BOLD_CYAN)Testing with pytest$(RESET_STYLES) 📊️"
 	pytest --version
 	pytest
 	@echo "\n"
