+++ requirements/dev-requirements.in
@@ -33,3 +33,6 @@ pre-commit
 
 # The pytest framework makes it easy to write small tests, yet scales to support complex functional testing for applications and libraries.
 pytest
+
+# This plugin produces coverage reports. Compared to just using coverage run, this plugin does some extras
+pytest-cov
