+++ setup.cfg
@@ -34,6 +34,7 @@ use_parentheses=True
 
 [tool:pytest]
 testpaths = test/pytest
+addopts = --cov=src --cov-report html
 
 [coverage:run]
 branch = True
