--- test/pytest/test_route_health.py
@@ -1,22 +1,0 @@
-from test.pytest.test_fixtures import TestFixtures
-
-from src.core import Logger
-from src.routes.health import Client
-
-log: Logger = Logger()
-
-
-class TestRouteHealth(TestFixtures):
-    @staticmethod
-    def url_prefix() -> str:
-        return Client.get_url_prefix()
-
-    # testing random path to give 404
-    def test_random(self, test_client) -> None:
-        health_response = test_client.get(self.url_prefix() + "/randomized")
-        assert health_response.status_code == 404
-
-    # testing ping path to give 200
-    def test_health(self, test_client) -> None:
-        health_response = test_client.get(self.url_prefix() + "/ping")
-        assert health_response.status_code == 200
